+++ pyanalyze/test_name_check_visitor.py
@@ -219,6 +219,8 @@ class TestNameCheckVisitor(TestNameCheckVisitorBase):
             class Capybara(metaclass=Hutia):  # E: undefined_name
                 pass
 
+            return Capybara
+
     @assert_passes()
     def test_no_failure_on_builtin(self):
         def run():
@@ -259,6 +261,8 @@ class TestNameCheckVisitor(TestNameCheckVisitorBase):
 
                 def coendou(self):
                     return 1
+
+            return Porcupine
 
     @assert_passes()
     def test_class_scope_is_not_searched(self):
@@ -414,6 +418,8 @@ class TestNameCheckVisitor(TestNameCheckVisitorBase):
             class Capybaras(object):
                 if False:
                     print(neochoerus)  # E: undefined_name
+
+            return Capybaras
 
     @assert_passes()
     def test_cant_del_tuple(self):
@@ -1005,6 +1011,8 @@ class TestComprehensions(TestNameCheckVisitorBase):
                 incisors = [1, 2]
                 canines = {incisors[0] for _ in incisors}  # E: undefined_name
 
+            return Capybara
+
     @assert_passes()
     def test_comprehension_within_class(self):
         class Capybara(object):
@@ -1386,7 +1394,7 @@ class TestNonlocal(TestNameCheckVisitorBase):
                 # this should not throw unused_variable
                 x = 5
 
-            return x
+            return x, inner_capybara, second_inner
 
     @assert_passes()
     def test_no_unused_var(self):
@@ -1397,7 +1405,7 @@ class TestNonlocal(TestNameCheckVisitorBase):
                 nonlocal running
                 running = False
 
-            return running
+            return running, handler
 
 
 class TestingCallSiteCollector(object):
@@ -1839,10 +1847,12 @@ class TestRequireAnnotations(TestNameCheckVisitorBase):
 
     @assert_passes(settings=_AnnotSettings)
     def test_dont_annotate_self():
-        def f() -> None:
+        def f() -> type:
             class X:
                 def method(self) -> None:
                     pass
+
+            return X
 
         class X:
             def f(self) -> None: