+++ pyanalyze/test_typeshed.py
@@ -259,6 +259,8 @@ class TestBundledStubs(TestNameCheckVisitorBase):
                 assert_is_value(pep655, _EXPECTED_TYPED_DICTS["PEP655"])
                 assert_is_value(inherited, _EXPECTED_TYPED_DICTS["Inherited"])
 
+            return nested
+
     def test_evaluated(self):
         tsf = TypeshedFinder.make(Checker(), TEST_OPTIONS, verbose=True)
         mod = "_pyanalyze_tests.evaluated"
@@ -303,6 +305,8 @@ class TestBundledStubs(TestNameCheckVisitorBase):
             def f(x: _ScandirIterator):
                 want_cm(x)
                 len(x)  # E: incompatible_argument
+
+            return f
 
     @assert_passes()
     def test_args_kwargs(self):